Rename ast.py to batch.py
