+++ src/snowflake/snowpark/_internal/ast/batch.py
